% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fknn.R
\name{fknn}
\alias{fknn}
\title{Calculate Fuzzy k-Nearest Neighbors}
\usage{
fknn(train_features, train_labels, test_features, k, m)
}
\description{
Calculate Fuzzy k-Nearest Neighbors
}
\details{
Keller, J. et al. “A fuzzy K-nearest neighbor algorithm.” IEEE Transactions on Systems, Man, and Cybernetics SMC-15 (1985): 580-585.
}
